+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "axios": "^1.5.1",
     "next": "13.5.5",
     "react": "^18",
     "react-dom": "^18"
