+++ src/app/axiosConfig.ts
@@ -1,0 +1,7 @@
+import axios from 'axios';
+
+const instance = axios.create({
+  baseURL: 'http://localhost:3333',
+});
+
+export default instance;
