+++ src/app/components/home/home.scss
@@ -20,10 +20,11 @@ html,body {
         justify-content: center;
         input{
           width: 75%;
-          margin-top: 2em;
+          height: 1.5em;
+          margin-top: 1em;
         }
         button{
-          margin-top: 2em;
+          margin-top: 1em;
           width: 75%;
         }
       }
